Fix the abort and visual glitches when 3D layers rotate through the camera near plane.

+++ src/layers/RootLayer.h
@@ -82,7 +82,7 @@ class RootLayer : public Layer {
 
  private:
   std::vector<Rect> dirtyRects = {};
-  std::vector<float> dirtyAreas = {};
+  std::vector<double> dirtyAreas = {};
 
   bool mergeDirtyList(bool forceMerge);
